Move around some of the PIC/IRQ functions + universal function to map IRQ to handler depending if using APIC or not
