+++ src/arch/x86_64/irq.h
@@ -7,10 +7,11 @@
 
 typedef void (*irq_handler)(registers_t *regs);
 
+void irq_registerHandler(int irq, irq_handler handler);
+
 extern void _enable_interrupts();
 extern void _disable_interrupts();
 
 void irq_init();
-void irq_registerHandler(int irq, irq_handler handler);
 
 #endif
